Update url and API fields name
